Implement autodetetion for self-signed certificate flag (#301)
Implement autodecetion for self-signed certificate flag
