+++ olm/eclipse-che-preview-kubernetes/deploy/olm-catalog/eclipse-che-preview-kubernetes/9.9.9-nightly.1592203276/eclipse-che-preview-kubernetes.v9.9.9-nightly.1592203276.clusterserviceversion.yaml.diff
@@ -1,0 +1,25 @@
+--- /home/mmorhun/projects/go/src/github.com/eclipse/che-operator/olm/eclipse-che-preview-kubernetes/deploy/olm-catalog/eclipse-che-preview-kubernetes/9.9.9-nightly.1591794793/eclipse-che-preview-kubernetes.v9.9.9-nightly.1591794793.clusterserviceversion.yaml	2020-06-15 09:11:07.460021934 +0300
++++ /home/mmorhun/projects/go/src/github.com/eclipse/che-operator/olm/eclipse-che-preview-kubernetes/deploy/olm-catalog/eclipse-che-preview-kubernetes/9.9.9-nightly.1592203276/eclipse-che-preview-kubernetes.v9.9.9-nightly.1592203276.clusterserviceversion.yaml	2020-06-15 09:41:17.218170783 +0300
+@@ -49,12 +49,12 @@
+     categories: Developer Tools
+     certified: "false"
+     containerImage: quay.io/eclipse/che-operator:nightly
+-    createdAt: "2020-06-10T13:13:13Z"
++    createdAt: "2020-06-15T06:41:17Z"
+     description: A Kube-native development solution that delivers portable and collaborative
+       developer workspaces.
+     repository: https://github.com/eclipse/che-operator
+     support: Eclipse Foundation
+-  name: eclipse-che-preview-kubernetes.v9.9.9-nightly.1591794793
++  name: eclipse-che-preview-kubernetes.v9.9.9-nightly.1592203276
+   namespace: placeholder
+ spec:
+   apiservicedefinitions: {}
+@@ -394,5 +394,5 @@
+   maturity: stable
+   provider:
+     name: Eclipse Foundation
+-  replaces: eclipse-che-preview-kubernetes.v9.9.9-nightly.1591355178
+-  version: 9.9.9-nightly.1591794793
++  replaces: eclipse-che-preview-kubernetes.v9.9.9-nightly.1591794793
++  version: 9.9.9-nightly.1592203276
