+++ olm/eclipse-che-preview-kubernetes/deploy/olm-catalog/eclipse-che-preview-kubernetes/eclipse-che-preview-kubernetes.package.yaml
@@ -1,5 +1,5 @@
 channels:
-- currentCSV: eclipse-che-preview-kubernetes.v9.9.9-nightly.1591794793
+- currentCSV: eclipse-che-preview-kubernetes.v9.9.9-nightly.1592203276
   name: nightly
 - currentCSV: eclipse-che-preview-kubernetes.v7.14.2
   name: stable
