+++ olm/eclipse-che-preview-openshift/deploy/olm-catalog/eclipse-che-preview-openshift/9.9.9-nightly.1592203277/eclipse-che-preview-openshift.v9.9.9-nightly.1592203277.clusterserviceversion.yaml.diff
@@ -1,0 +1,25 @@
+--- /home/mmorhun/projects/go/src/github.com/eclipse/che-operator/olm/eclipse-che-preview-openshift/deploy/olm-catalog/eclipse-che-preview-openshift/9.9.9-nightly.1591794794/eclipse-che-preview-openshift.v9.9.9-nightly.1591794794.clusterserviceversion.yaml	2020-06-15 09:11:07.463021953 +0300
++++ /home/mmorhun/projects/go/src/github.com/eclipse/che-operator/olm/eclipse-che-preview-openshift/deploy/olm-catalog/eclipse-che-preview-openshift/9.9.9-nightly.1592203277/eclipse-che-preview-openshift.v9.9.9-nightly.1592203277.clusterserviceversion.yaml	2020-06-15 09:41:18.389177661 +0300
+@@ -46,12 +46,12 @@
+     categories: Developer Tools, OpenShift Optional
+     certified: "false"
+     containerImage: quay.io/eclipse/che-operator:nightly
+-    createdAt: "2020-06-10T13:13:14Z"
++    createdAt: "2020-06-15T06:41:17Z"
+     description: A Kube-native development solution that delivers portable and collaborative
+       developer workspaces in OpenShift.
+     repository: https://github.com/eclipse/che-operator
+     support: Eclipse Foundation
+-  name: eclipse-che-preview-openshift.v9.9.9-nightly.1591794794
++  name: eclipse-che-preview-openshift.v9.9.9-nightly.1592203277
+   namespace: placeholder
+ spec:
+   apiservicedefinitions: {}
+@@ -438,5 +438,5 @@
+   maturity: stable
+   provider:
+     name: Eclipse Foundation
+-  replaces: eclipse-che-preview-openshift.v9.9.9-nightly.1591355180
+-  version: 9.9.9-nightly.1591794794
++  replaces: eclipse-che-preview-openshift.v9.9.9-nightly.1591794794
++  version: 9.9.9-nightly.1592203277
